parse API: make bytes vs characters explicit
Previously we returned `start` and `end` for each Range and the values were bytes, not chars. This is easy to miss and leads to bugs when you try to handle strings with multi-byte characters.

This breaking change renames `start` to `startByte` and `end` to `endByte` and also adds `startChar` and `endChar` so that users of the `parse()` API can pick the one they really want.